add proposalbyusername metod

+++ ScoreMe.DAL/Objects/Search.cs
@@ -15,7 +15,7 @@ namespace ScoreMe.DAL.Objects
         public Int64 Id { get; set; }
         public Int64 ProviderID{ get; set; }
         public Int64 ProposalID { get; set; }
-        public int UserId { get; set; }
+        public Int64 UserId { get; set; }
         public string UserName { get; set; }
         public bool isCount { get; set; }
